+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Guide: https://keepachangelog.com/en/1.0.0/
 
 - [Tests] Change MockResponse into a protocol, create separate enums conforming to MockResponse for each API type (geocoding, sbs, autofill), add MockResponse as generic to each test base class and MockWebServer.
 
+- [Discover] Add support for country, proximity, and origin parameters in Discover.Options search parameters. This fixes an issue when using search-along-route to query category results.
+
 - [SearchUI] Add `distanceFormatter` field to Configuration to support changing the search suggestions distance format. Nil values will use the default behavior.
 
 - [Core] Add xcprivacy for MapboxSearch and MapboxSearchUI